+++ CHANGELOG.md
@@ -5,6 +5,8 @@ to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+- Added Create Release Command (Alpha)
+
 ## [1.1.0] - 2021-06-29
 ### Added
 - DefaultChecker
@@ -34,7 +36,7 @@ to [Semantic Versioning](http://semver.org/).
 ### Added
 - Create Project
 
-[Unreleased]: https://github.com/boscho87/changelog-checker/compare/0.0.4...master
+[Unreleased]: https://github.com/boscho87/changelog-checker/compare/1.1.0...master
 [1.1.0]: https://github.com/boscho87/changelog-checker/compare/1.0.0...1.1.0
 [1.0.0]: https://github.com/boscho87/changelog-checker/compare/0.0.4...1.0.0
 [0.0.4]: https://github.com/boscho87/changelog-checker/compare/0.0.3...0.0.4